Fixed bug in loop initialization.
Also fixed bug with OcmController::maxH|VFrac member (was int32_t, now a double).

Bug:
double max;
for(int i=0,max=0, ...) {...}

Soln:
double max=0;
for(int i=0;...) {...}

In the first instance there are two local \"max\" vars created: one dbl and one int.

+++ rtems/include/OrbitController.h
@@ -174,8 +174,8 @@ private:
 	};
 	int32_t maxHStep;
 	int32_t maxVStep;
-	int32_t maxHFrac;
-	int32_t maxVFrac;
+	double maxHFrac;
+	double maxVFrac;
 	/* FIXME -- replace static arrays with vector<vector<double>> && vector<double>
 	 * NOTE  -- also replace const NumOcm/NumBpm,etc with static variables and let
 	 * 			the UI inform *us* of how many OCM/BPM are required...
